requirements.txt
Django==3.0
django-countries==5.5
envdir==1.0.1
djangorestframework==3.11.0
psycopg2-binary==2.8.4